+++ GRDB Tests/Public/Foundation/RowConvertible+FoundationTests.swift
@@ -15,11 +15,17 @@ struct PointOfInterest : RowConvertible {
 
 class RowConvertibleFoundationTests: GRDBTestCase {
     
+    func testInvalidNSDictionaryInitializer() {
+        let dictionary: NSDictionary = ["a": NSObject()]
+        let s = PointOfInterest(dictionary: dictionary)
+        XCTAssertTrue(s == nil)
+    }
+    
     func testNSDictionaryInitializer() {
         let latitude = 41.8919300
         let longitude = 12.5113300
         let dictionary: NSDictionary = ["latitude": latitude, "longitude": longitude, "title": NSNull()]
-        let s = PointOfInterest(dictionary: dictionary)
+        let s = PointOfInterest(dictionary: dictionary)!
         XCTAssertEqual(s.latitude, latitude)
         XCTAssertEqual(s.longitude, longitude)
         XCTAssertTrue(s.title == nil)